Refacto userWithPerimeter object in user services
