+++ backend/cards-consultation/src/test/java/org/opfab/cards/consultation/routes/ConnectionRoutesShould.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2021-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -141,9 +141,9 @@ class ConnectionRoutesShould {
             service.subscribe(createUserWithPerimeter("testuser3"), "test3").getPublisher().subscribe(log::info);
 
             String[] expectedUsers = {
-                    "{\"login\":\"testuser\",\"firstName\":null,\"lastName\":null,\"entitiesConnected\":null,\"groups\":null}",
-                    "{\"login\":\"testuser2\",\"firstName\":null,\"lastName\":null,\"entitiesConnected\":null,\"groups\":null}",
-                    "{\"login\":\"testuser3\",\"firstName\":null,\"lastName\":null,\"entitiesConnected\":null,\"groups\":null}" };
+                    "{\"login\":\"testuser\",\"firstName\":null,\"lastName\":null,\"entitiesConnected\":[],\"groups\":[]}",
+                    "{\"login\":\"testuser2\",\"firstName\":null,\"lastName\":null,\"entitiesConnected\":[],\"groups\":[]}",
+                    "{\"login\":\"testuser3\",\"firstName\":null,\"lastName\":null,\"entitiesConnected\":[],\"groups\":[]}" };
 
             webTestClient.get().uri("/connections").exchange().expectStatus().isOk();
