+++ backend/users/src/main/java/org/opfab/users/configuration/authorization/UserExtractor.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2018-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -9,9 +9,8 @@
 
 package org.opfab.users.configuration.authorization;
 
-import org.opfab.users.configuration.jwt.OpFabJwtAuthenticationToken;
-import org.opfab.users.model.User;
-
+import org.opfab.configuration.oauth.OpFabJwtAuthenticationToken;
+import org.opfab.common.users.User;
 import jakarta.servlet.http.HttpServletRequest;
 import java.security.Principal;
