--- backend/users/src/main/java/org/opfab/users/configuration/jwt/OpFabJwtAuthenticationToken.java
@@ -1,70 +1,0 @@
-/* Copyright (c) 2024-2025, RTE (http://www.rte-france.com)
- * See AUTHORS.txt
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- * SPDX-License-Identifier: MPL-2.0
- * This file is part of the OperatorFabric project.
- */
-
-package org.opfab.users.configuration.jwt;
-
-import org.opfab.users.model.CurrentUserWithPerimeters;
-import org.opfab.users.model.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
-
-import java.util.Collection;
-
-/**
- * See README.md for documentation
- */
-public class OpFabJwtAuthenticationToken extends JwtAuthenticationToken {
-
-    private static final long serialVersionUID = 1L;
-
-    private final transient Object principal;
-
-    public OpFabJwtAuthenticationToken(Jwt jwt, User principal, Collection<? extends GrantedAuthority> authorities) {
-        super(jwt, authorities);
-        this.principal = principal;
-    }
-
-    public OpFabJwtAuthenticationToken(Jwt jwt, CurrentUserWithPerimeters principal,
-            Collection<? extends GrantedAuthority> authorities) {
-        super(jwt, authorities);
-        this.principal = principal;
-    }
-
-    @Override
-    public Object getPrincipal() {
-        return this.principal;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = super.hashCode();
-        result = prime * result + ((principal == null) ? 0 : principal.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (!super.equals(obj))
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        OpFabJwtAuthenticationToken other = (OpFabJwtAuthenticationToken) obj;
-        if (principal == null) {
-            if (other.principal != null)
-                return false;
-        } else if (!principal.equals(other.principal))
-            return false;
-        return true;
-    }
-
-}
