+++ backend/users/src/main/java/org/opfab/users/configuration/users/UsersProperties.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2018-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -9,7 +9,11 @@
 
 package org.opfab.users.configuration.users;
 
-import org.opfab.users.model.*;
+import org.opfab.common.users.User;
+import org.opfab.users.model.Entity;
+import org.opfab.users.model.Group;
+import org.opfab.users.model.Perimeter;
+import org.opfab.users.model.UserSettings;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
