+++ backend/users/src/main/java/org/opfab/users/controllers/UserWithPerimetersController.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -12,7 +12,6 @@ package org.opfab.users.controllers;
 import org.opfab.useractiontracing.repositories.LastUserActionRepository;
 import org.opfab.useractiontracing.repositories.UserActionLogRepository;
 import org.opfab.useractiontracing.services.UserActionLogService;
-import org.opfab.users.model.*;
 import org.opfab.utilities.eventbus.EventBus;
 import org.opfab.users.repositories.EntityRepository;
 import org.opfab.users.repositories.GroupRepository;
@@ -23,6 +22,7 @@ import org.opfab.users.services.CurrentUserWithPerimetersService;
 import org.opfab.users.services.NotificationService;
 import org.opfab.users.services.UserSettingsService;
 import org.opfab.users.services.UsersService;
+import org.opfab.common.users.CurrentUserWithPerimeters;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -40,21 +40,22 @@ public class UserWithPerimetersController {
     private CurrentUserWithPerimetersService currentUserWithPerimetersService;
 
     public UserWithPerimetersController(UserRepository userRepository, GroupRepository groupRepository,
-                                        PerimeterRepository perimeterRepository, EntityRepository entityRepository,
-                                        UserSettingsRepository userSettingsRepository, UserActionLogRepository userActionLogRepository,
-                                        EventBus eventBus,
-                                        @Value("${operatorfabric.userActionLogActivated:true}") boolean userActionLogActivated,
-                                        LastUserActionRepository lastUserActionRepository) {
+            PerimeterRepository perimeterRepository, EntityRepository entityRepository,
+            UserSettingsRepository userSettingsRepository, UserActionLogRepository userActionLogRepository,
+            EventBus eventBus,
+            @Value("${operatorfabric.userActionLogActivated:true}") boolean userActionLogActivated,
+            LastUserActionRepository lastUserActionRepository) {
 
-            NotificationService notificationService = new NotificationService(userRepository, eventBus);
-            UsersService usersService = new UsersService(userRepository, groupRepository, entityRepository,
-                            perimeterRepository, notificationService);
-            UserActionLogService userActionLogService = new UserActionLogService(userActionLogRepository, lastUserActionRepository);
-            UserSettingsService userSettingsService = new UserSettingsService(userSettingsRepository, usersService,
-                            notificationService, userActionLogService, userActionLogActivated);
-            this.currentUserWithPerimetersService = new CurrentUserWithPerimetersService(usersService,
-                            userSettingsService,
-                            entityRepository);
+        NotificationService notificationService = new NotificationService(userRepository, eventBus);
+        UsersService usersService = new UsersService(userRepository, groupRepository, entityRepository,
+                perimeterRepository, notificationService);
+        UserActionLogService userActionLogService = new UserActionLogService(userActionLogRepository,
+                lastUserActionRepository);
+        UserSettingsService userSettingsService = new UserSettingsService(userSettingsRepository, usersService,
+                notificationService, userActionLogService, userActionLogActivated);
+        this.currentUserWithPerimetersService = new CurrentUserWithPerimetersService(usersService,
+                userSettingsService,
+                entityRepository);
     }
 
     @GetMapping(value = "/{login}", produces = { "application/json" })
