+++ backend/users/src/main/java/org/opfab/users/model/StateRight.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2018-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,25 +7,23 @@
  * This file is part of the OperatorFabric project.
  */
 
-
-
 package org.opfab.users.model;
 
 import org.springframework.validation.annotation.Validated;
-
 import com.fasterxml.jackson.annotation.JsonInclude;
 
+import org.opfab.common.users.RightEnum;
 
 @Validated
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class StateRight{
+public class StateRight {
     private String state;
     private RightEnum right;
     private Boolean filteringNotificationAllowed = true;
 
     public StateRight() {
     }
-    
+
     public StateRight(String state, RightEnum right, Boolean filteringNotificationAllowed) {
         this.state = state;
         this.right = right;
@@ -39,7 +37,7 @@ public class StateRight{
     public void setState(String state) {
         this.state = state;
     }
-    
+
     public RightEnum getRight() {
         return right;
     }
@@ -61,8 +59,6 @@ public class StateRight{
         return "StateRightData [state=" + state + ", right=" + right + ", filteringNotificationAllowed="
                 + filteringNotificationAllowed + "]";
     }
-
-    
 
     @Override
     public int hashCode() {
@@ -99,5 +95,4 @@ public class StateRight{
         return true;
     }
 
-    
 }
