+++ backend/users/src/main/java/org/opfab/users/mongo/repositories/MongoUserRepository.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,10 +7,9 @@
  * This file is part of the OperatorFabric project.
  */
 
-
 package org.opfab.users.mongo.repositories;
 
-import org.opfab.users.model.User;
+import org.opfab.common.users.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -23,10 +22,13 @@ import java.util.Optional;
  * Mongo {@link User} repository
  */
 @Repository
-public interface MongoUserRepository extends MongoRepository<User,String> {
+public interface MongoUserRepository extends MongoRepository<User, String> {
 
     Page<User> findAll(Pageable pageable);
+
     Optional<User> findByLogin(String login);
+
     List<User> findByGroupSetContaining(String groupContains);
+
     List<User> findByEntitiesContaining(String entityContains);
 }
