+++ backend/users/src/main/java/org/opfab/users/mongo/repositories/MongoUserSettingsRepository.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,10 +7,9 @@
  * This file is part of the OperatorFabric project.
  */
 
-
 package org.opfab.users.mongo.repositories;
 
-import org.opfab.users.model.User;
+import org.opfab.common.users.User;
 import org.opfab.users.model.UserSettings;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
@@ -20,6 +19,6 @@ import org.springframework.stereotype.Repository;
  *
  */
 @Repository
-public interface MongoUserSettingsRepository extends MongoRepository<UserSettings,String> {
+public interface MongoUserSettingsRepository extends MongoRepository<UserSettings, String> {
 
 }
