+++ backend/users/src/main/java/org/opfab/users/mongo/repositories/UserRepositoryImpl.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -13,7 +13,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import org.opfab.users.model.User;
+import org.opfab.common.users.User;
 import org.opfab.users.repositories.UserRepository;
 
 public class UserRepositoryImpl implements UserRepository {
@@ -67,7 +67,8 @@ public class UserRepositoryImpl implements UserRepository {
 
     @Override
     public void delete(User user) {
-        if (user == null) return;
+        if (user == null)
+            return;
         mongoUserRepository.delete(user);
 
     }
