+++ backend/users/src/main/java/org/opfab/users/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,34 +7,33 @@
  * This file is part of the OperatorFabric project.
  */
 
-
 package org.opfab.users.repositories;
 
 import java.util.List;
 import java.util.Optional;
 
-import org.opfab.users.model.User;
+import org.opfab.common.users.User;
 
 public interface UserRepository {
 
-    public List<User>  saveAll(List<User>  users);
+    public List<User> saveAll(List<User> users);
 
-    public List<User>  findAll();
+    public List<User> findAll();
 
     public User insert(User user);
 
     public User save(User user);
 
-    public Optional<User>  findById(String id);
+    public Optional<User> findById(String id);
 
-    public Optional<User>  findByLogin(String userId);
+    public Optional<User> findByLogin(String userId);
 
     public void delete(User user);
 
     public void deleteAll();
 
-    public List<User>  findByGroupSetContaining(String groupContains);
+    public List<User> findByGroupSetContaining(String groupContains);
 
-    public List<User>  findByEntitiesContaining(String entityContains);
+    public List<User> findByEntitiesContaining(String entityContains);
 
 }
