+++ backend/users/src/test/java/org/opfab/users/services/UserSettingsServiceShould.java
@@ -1,5 +1,5 @@
 
-/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -24,9 +24,9 @@ import org.junit.jupiter.api.Test;
 import org.opfab.test.EventBusSpy;
 import org.opfab.users.model.OperationResult;
 import org.opfab.users.model.Perimeter;
-import org.opfab.users.model.RightEnum;
+import org.opfab.common.users.RightEnum;
 import org.opfab.users.model.StateRight;
-import org.opfab.users.model.User;
+import org.opfab.common.users.User;
 import org.opfab.users.model.UserSettings;
 import org.opfab.users.stubs.UserRepositoryStub;
 import org.opfab.users.stubs.UsersServiceStub;
