+++ backend/users/src/test/java/org/opfab/users/services/UsersServiceShould.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -25,9 +25,9 @@ import org.opfab.users.model.Entity;
 import org.opfab.users.model.Group;
 import org.opfab.users.model.OperationResult;
 import org.opfab.users.model.Perimeter;
-import org.opfab.users.model.RightEnum;
+import org.opfab.common.users.RightEnum;
 import org.opfab.users.model.StateRight;
-import org.opfab.users.model.User;
+import org.opfab.common.users.User;
 import org.opfab.users.stubs.EntityRepositoryStub;
 import org.opfab.users.stubs.GroupRepositoryStub;
 import org.opfab.users.stubs.PerimeterRepositoryStub;
