+++ frontend/src/app/components/dashboard/view/DashboardView.ts
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2026, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -180,7 +180,7 @@ export class Dashboard {
     }
 
     private isStateNotified(id: string, name: string): boolean {
-        if (UsersService.getCurrentUserWithPerimeters().processesStatesNotNotified.has(id)) {
+        if (UsersService.getCurrentUserWithPerimeters().processesStatesNotNotified?.has(id)) {
             return UsersService.getCurrentUserWithPerimeters().processesStatesNotNotified.get(id).indexOf(name) <= -1;
         }
         return true;
